+++ store/plugins/websocket.js
@@ -20,8 +20,12 @@ export default function create_web_socket_plugin(socket) {
       }
     });
     socket.on("twitch_metrics", function (metrics_json, cb = null) {
-      // guard against metrics coming right after live view stops so heatmap stay cleared
-      if (store.state.playback.playback_state !== ENUMS.PLAYBACK_STATES.LIVE_VIEW_ACTIVE) {
+      // guard against metrics coming right after live view stops so heatmap stays cleared,
+      // also need to make sure heatmap can update while recording
+      if (
+        store.state.playback.playback_state !== ENUMS.PLAYBACK_STATES.LIVE_VIEW_ACTIVE &&
+        store.state.playback.playback_state !== ENUMS.PLAYBACK_STATES.RECORDING
+      ) {
         return;
       }
       store.commit("data/append_metric_data", JSON.parse(metrics_json));